third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/insecure-image-in-iframe-expected.txt

CONSOLE WARNING: line 9: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image.html' was loaded over HTTPS, but requested an insecure image 'http://example.test:8080/security/resources/compass.jpg'. This content should also be served over HTTPS.
This test loads a secure iframe that loads an insecure image. We should get a mixed content callback becase the frame is HTTPS and the image is not.
